Few Business logics moved to service layer
