+++ src/main/java/com/searchservice/app/rest/InputDocumentResource.java
@@ -49,10 +49,11 @@ public class InputDocumentResource {
 
 	}
 
+	
 	@RateLimiter(name = DOCUMENT_INJECTION_THROTTLER_SERVICE, fallbackMethod = "documentInjectionRateLimiterFallback")
 	@PostMapping("/ingest-nrt/{tableName}")
 	@Operation(summary = "ADD DOCUMENTS IN THE TABLE OF THE GIVEN TENANT ID. INPUT SHOULD BE A LIST OF DOCUMENTS SATISFYING THE TABLE SCHEMA. NEAR REAL-TIME API.", security = @SecurityRequirement(name = "bearerAuth"))
-	public ResponseEntity<ThrottlerResponse> addDocumentsWithNRT(@RequestParam int tenantId, @PathVariable String tableName,
+	public ResponseEntity<ThrottlerResponse> addDocumentsNRT(@RequestParam int tenantId, @PathVariable String tableName,
 			@RequestBody String payload) {
 
 		if (!inputDocumentService.isValidJsonArray(payload))
@@ -69,17 +70,16 @@ public class InputDocumentResource {
 		// Control will reach here ONLY IF REQUESTBODY SIZE IS UNDER THE SPECIFIED LIMIT
 		tableName = tableName + "_" + tenantId;
 		if (manageTableServicePort.isTableExists(tableName)) {
-
-			return performDocumentInjection(true,tableName, payload, documentInjectionThrottlerResponse);
+			return inputDocumentServicePort.performDocumentInjection(true,tableName, payload, documentInjectionThrottlerResponse);
 		} else {
-			return documentInjectWithInvalidTableName(tenantId, tableName.split("_")[0]);
+			return inputDocumentServicePort.documentInjectWithInvalidTableName(tenantId, tableName.split("_")[0]);
 		}
 	}
 
 	@RateLimiter(name = DOCUMENT_INJECTION_THROTTLER_SERVICE, fallbackMethod = "documentInjectionRateLimiterFallback")
 	@PostMapping("/ingest/{tableName}")
 	@Operation(summary = "ADD DOCUMENTS IN THE TABLE OF THE GIVEN TENANT ID. INPUT SHOULD BE A LIST OF DOCUMENTS SATISFYING THE TABLE SCHEMA.", security = @SecurityRequirement(name = "bearerAuth"))
-	public ResponseEntity<ThrottlerResponse> addDocumentWithoutNRT(@RequestParam int tenantId, @PathVariable String tableName,
+	public ResponseEntity<ThrottlerResponse> addDocumentsBatch(@RequestParam int tenantId, @PathVariable String tableName,
 			@RequestBody String payload) {
 
 		if (!inputDocumentService.isValidJsonArray(payload))
@@ -100,9 +100,9 @@ public class InputDocumentResource {
 
 		// Control will reach here ONLY IF REQUESTBODY SIZE IS UNDER THE SPECIFIED LIMIT
 		if (manageTableServicePort.isTableExists(tableName)) {
-			return performDocumentInjection(false, tableName, payload, documentInjectionThrottlerResponse);
+			return inputDocumentServicePort.performDocumentInjection(false, tableName, payload, documentInjectionThrottlerResponse);
 		} else {
-			return documentInjectWithInvalidTableName(tenantId, tableName.split("_")[0]);
+			return inputDocumentServicePort.documentInjectWithInvalidTableName(tenantId, tableName.split("_")[0]);
 		}
 	}
 
@@ -121,25 +121,4 @@ public class InputDocumentResource {
 				.body(rateLimitResponseDTO);
 	}
 
-	public ResponseEntity<ThrottlerResponse> performDocumentInjection(boolean isNrt, String tableName, String payload,
-			ThrottlerResponse documentInjectionThrottlerResponse) {
-		ThrottlerResponse documentInjectionResponse = inputDocumentServicePort.addDocuments(isNrt, tableName, payload);
-
-		documentInjectionThrottlerResponse.setMessage(documentInjectionResponse.getMessage());
-		documentInjectionThrottlerResponse.setStatusCode(documentInjectionResponse.getStatusCode());
-		if (documentInjectionThrottlerResponse.getStatusCode() == 200) {
-
-			return ResponseEntity.status(HttpStatus.OK).body(documentInjectionThrottlerResponse);
-		} else {
-
-			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(documentInjectionThrottlerResponse);
-		}
-	}
-
-	public ResponseEntity<ThrottlerResponse> documentInjectWithInvalidTableName(int tenantId, String tableName){
-		ThrottlerResponse documentInjectionThrottlerResponse= new ThrottlerResponse();
-		documentInjectionThrottlerResponse.setStatusCode(HttpStatusCode.TABLE_NOT_FOUND.getCode());
-    	documentInjectionThrottlerResponse.setMessage("Table "+tableName+" For Tenant ID: "+tenantId+ " "+ HttpStatusCode.TABLE_NOT_FOUND.getMessage());
-    	return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(documentInjectionThrottlerResponse);
-	}
 }
